refactor: Money 타입 long 으로 변경

+++ src/main/java/kr/hhplus/be/server/application/balance/BalanceInfo.java
@@ -2,12 +2,11 @@ package kr.hhplus.be.server.application.balance;
 
 import kr.hhplus.be.server.domain.balance.Balance;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record BalanceInfo(
         Long userId,
-        BigDecimal amount,
+        Long amount,
         LocalDateTime updatedAt
 ) {
     public static BalanceInfo from(Balance balance) {
