+++ src/main/java/kr/hhplus/be/server/application/balance/ChargeBalanceCommand.java
@@ -1,8 +1,7 @@
 package kr.hhplus.be.server.application.balance;
 
-import java.math.BigDecimal;
 
 public record ChargeBalanceCommand(
         Long userId,
-        BigDecimal amount
+        long amount
 ) {}
