+++ src/main/java/kr/hhplus/be/server/application/balance/ChargeBalanceCriteria.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public record ChargeBalanceCriteria(
         Long userId,
-        BigDecimal amount,
+        long amount,
         String reason
 ) {}
