+++ src/main/java/kr/hhplus/be/server/application/balance/RecordBalanceHistoryCommand.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 public record RecordBalanceHistoryCommand(
         Long userId,
-        BigDecimal amount,
+        long amount,
         BalanceChangeType type,
         String reason
 ) {}
