+++ src/main/java/kr/hhplus/be/server/domain/balance/BalanceHistory.java
@@ -22,7 +22,7 @@ public class BalanceHistory {
     private Long userId;
 
     @Column(nullable = false)
-    private BigDecimal amount;
+    private long amount;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
@@ -33,7 +33,7 @@ public class BalanceHistory {
     @Column(nullable = false)
     private LocalDateTime createdAt;
 
-    public BalanceHistory(Long userId, BigDecimal amount, BalanceChangeType type, String reason, LocalDateTime createdAt) {
+    public BalanceHistory(Long userId, long amount, BalanceChangeType type, String reason, LocalDateTime createdAt) {
         this.userId = userId;
         this.amount = amount;
         this.type = type;
@@ -41,7 +41,7 @@ public class BalanceHistory {
         this.createdAt = createdAt;
     }
 
-    public static BalanceHistory of(Long userId, BigDecimal amount, BalanceChangeType type, String reason) {
+    public static BalanceHistory of(Long userId, long amount, BalanceChangeType type, String reason) {
         return new BalanceHistory(userId, amount, type, reason, LocalDateTime.now());
     }
