+++ src/main/java/kr/hhplus/be/server/interfaces/balance/BalanceResponse.java
@@ -13,12 +13,12 @@ public class BalanceResponse {
     private final Long userId;
 
     @Schema(description = "현재 잔액", example = "150000")
-    private final BigDecimal balance;
+    private final long balance;
 
     @Schema(description = "잔액 갱신 시각", example = "2025-04-02T10:30:00")
     private final LocalDateTime updatedAt;
 
-    public BalanceResponse(Long userId, BigDecimal balance, LocalDateTime updatedAt) {
+    public BalanceResponse(Long userId, long balance, LocalDateTime updatedAt) {
         this.userId = userId;
         this.balance = balance;
         this.updatedAt = updatedAt;
